+++ src/App.js
@@ -1,16 +1,17 @@
 import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import ProducDetails from './pages/ProductDetails';
+
+import ProductDetails from './pages/ProductDetails';
 import Home from './pages/Home';
-import Checkout from './pages/Checkout';
+import ShoppingCart from './pages/ShoppingCart';
 
 function App() {
   return (
     <BrowserRouter>
       <Switch>
         <Route exact path="/" component={ Home } />
-        <Route path="/checkout" component={ Checkout } />
-        <Route path="/product-details/:id" component={ ProducDetails } />
+        <Route path="/ShoppingCart" component={ ShoppingCart } />
+        <Route path="/product-details/:id" component={ ProductDetails } />
       </Switch>
     </BrowserRouter>
   );
