+++ src/pages/ShoppingCart.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import ProductCard from '../components/ProductCart';
+import ProductCard from '../components/ProductCard';
 
-class Checkout extends React.Component {
+class ShoppingCart extends React.Component {
   render() {
     const cartList = localStorage.getItem('shoppintCart');
     if (!cartList) {
@@ -18,4 +18,4 @@ class Checkout extends React.Component {
   }
 }
 
-export default Checkout;
+export default ShoppingCart;
